Module tinymt::tinymt32

Functions

tinymt32_init
This function initializes the internal state array with a 32-bit unsigned integer seed. @param random tinymt state vector. @param seed a 32-bit unsigned integer used as a seed.

tinymt32_init_by_array
This function initializes the internal state array, with an array of 32-bit unsigned integers used as seeds @param init_key the array of 32-bit integers, used as a seed. @param key_length the length of init_key.

tinymt32_get_mexp
This function always returns 127 @return always 127

tinymt32_next_state
This function changes internal state of tinymt32. Users should not call this function directly. @param random tinymt internal status

tinymt32_temper
This function outputs 32-bit unsigned integer from internal state. Users should not call this function directly. @param random tinymt internal status @return 32-bit unsigned pseudorandom number

tinymt32_temper_conv
This function outputs floating point number from internal state. Users should not call this function directly. @param random tinymt internal status @return floating point number r (1.0 <= r < 2.0)

tinymt32_temper_conv_open
This function outputs floating point number from internal state. Users should not call this function directly. @return floating point number r (1.0 < r < 2.0)

tinymt32_generate_float
This function outputs floating point number from internal state. This function is implemented using multiplying by (1 / 2^24). floating point multiplication is faster than using union trick in my Intel CPU. @return floating point number r (0.0 <= r < 1.0)

tinymt32_generate_float_oc
This function outputs floating point number from internal state. This function may return 1.0 and never returns 0.0. @return floating point number r (0.0 < r <= 1.0)

tinymt32_generate_float_oo
This function outputs floating point number from internal state. This function returns neither 0.0 nor 1.0. @return floating point number r (0.0 < r < 1.0)

tinymt32_generate_32double
This function outputs double precision floating point number from internal state. The returned value has 32-bit precision. In other words, this function makes one double precision floating point number from one 32-bit unsigned integer. @return floating point number r (0.0 <= r < 1.0)

tinymt32_generate_float01
This function outputs floating point number from internal state. This function is implemented using union trick. @return floating point number r (0.0 <= r < 1.0)

tinymt32_generate_float12
This function outputs floating point number from internal state. This function is implemented using union trick. @return floating point number r (1.0 <= r < 2.0)

tinymt32_generate_uint32
This function outputs 32-bit unsigned integer from internal state. @return 32-bit unsigned integer r (0 <= r < 2^32)